refactor styles and move security tab legacy settings to v2
