+++ ui/pages/settings-v2/security-and-password-tab/manage-wallet-recovery-sub-page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { RevealSrpList } from '../../settings/security-tab/reveal-srp-list';
+import { RevealSrpList } from './reveal-srp-list';
 
 const ManageWalletRecoverySubPage = () => {
   return <RevealSrpList />;
