+++ ui/pages/settings/index.scss
@@ -2,7 +2,6 @@
 
 @import '../settings-v2/debug-tab/developer-options-tab/index';
 @import 'networks-tab/index';
-@import 'security-tab/index';
 @import 'transaction-shield-tab/index';
 
 .settings-page {
