--- ui/pages/settings/security-tab/index.js
@@ -1,1 +1,0 @@
-export { default } from './security-tab.container';
