--- ui/pages/settings/security-tab/index.scss
@@ -1,50 +1,0 @@
-@use "design-system";
-
-.protect-wallet {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 1rem;
-
-  &__container {
-    border-radius: 0.5rem;
-    background-color: var(--color-background-muted);
-  }
-
-  &__container-body {
-    border-top: 1px solid var(--color-background-default);
-  }
-}
-
-.change-password {
-  &__spinner {
-    width: 1.5rem;
-    height: 1.5rem;
-    margin-bottom: 1rem;
-  }
-
-  &__link-text {
-    color: var(--color-primary-default);
-  }
-
-  @include design-system.screen-sm-max {
-    height: 100%;
-  }
-}
-
-.srp-reveal-list {
-  &__divider {
-    height: 1px;
-    background-color: var(--color-border-muted);
-  }
-
-  &__social-login-card {
-    padding: 12px 16px;
-    border-radius: 8px;
-  }
-
-  .select-srp__container {
-    background-color: var(--color-background-muted);
-    border: 0;
-  }
-}
